message

+++ post/views.py
@@ -10,7 +10,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 
 
 # Create your views here.
-class PostListView(View,LoginRequiredMixin):
+class PostListView(LoginRequiredMixin,View):
     def get(self,request, *args , **kwargs):
         posts = Post.objects.filter(user=request.user)
         paginator = Paginator(posts,8)
@@ -32,11 +32,19 @@ class PostDetailView(View):
         return render(request,"post/detail",context={'post':post})
     
     
-class PostEditView(View):
+class PostEditView(LoginRequiredMixin,View):
     def get(self,request,id,*args,**kwargs):
-        post = Post.objects.get(id=id)
-        form = PostForm(instance=post)
-        return render(request,"post/edit.html",context={'form':form})
+        try:
+            post = Post.objects.get(id=id)
+            if post.user!=request.user:
+                return redirect(reverse("403_error"))
+            form = PostForm(instance=post)
+            return render(request,"post/edit.html",context={'form':form})
+        except Post.DoesNotExist:
+            return redirect(reverse("404_Error"))
+        except Exception:
+            return redirect(reverse("500_Error"))
+
 
     def post(self,request,*args,**kwargs):
         post = Post.objects.get(id=id)
@@ -46,11 +54,10 @@ class PostEditView(View):
             return redirect(reverse("post-list"))
         return render(request,"post/edit.html",context={"form":form})
     
-class PostCreateView(View):
+class PostCreateView(LoginRequiredMixin,View):#only 403
     def get(self,request,*args,**kwargs):
         form = PostForm()
         return render(request,"post/create.html",context={'form':form})
-
     def post(self,request,*args,**kwargs):
         print("POST DATA:",request.POST,request.FILES)
         form = PostForm(request.POST, request.FILES)
@@ -60,9 +67,11 @@ class PostCreateView(View):
             return redirect(reverse('post-list'))
         return render (request,"post/create.html",context={'form':form})
     
-class PostDeleteView(View):
+class PostDeleteView(LoginRequiredMixin,View):
     def post(self,request,*args,**kwargs):
         post = Post.objects.get(id=request.POST.get("id"))
+        if post.user!=request.user:
+            return redirect(reverse("404_error"))
         post.delete()
         return redirect(reverse('post-list'))
